LayoutTests/http/tests/security/sandboxed-iframe-ALLOWED-modals.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();
</script>
<p>This test passes if opening modal dialogs is allowed and no error message is logged in the console.</p>
<iframe sandbox="allow-scripts allow-modals allow-same-origin" src="resources/sandboxed-iframe-ALLOWED-modals-iframe.html"></iframe>
