LayoutTests/http/tests/security/sandboxed-iframe-ALLOWED-modals-expected.txt

ALERT: MESSAGE
CONFIRM: MESSAGE?
PROMPT: MESSAGE:, default text: DEFAULT
This test passes if opening modal dialogs is allowed and no error message is logged in the console.
